+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Unreleased
 
+## 0.1.2
+
 ### Added
 
 - Checkout a different branch, tag or commit upon plugin install (`--branch`, `--tag`, `--commit` arguments for install command)